new changes

+++ Portal/src/app/user-master/user-master.component.ts
@@ -4,10 +4,14 @@ import { UserManagement } from '../Services/UserManagementService';
 import { Router } from '@angular/router';
 import { DialogService } from 'primeng/dynamicdialog';
 import { DialogeComponent } from '../dialoge/dialoge.component';
+import { ConfirmationService } from 'primeng/api';
+import { ConfirmDialog } from 'primeng/confirmdialog';
+
 @Component({
   selector: 'app-user-master',
   templateUrl: './user-master.component.html',
-  styleUrls: ['./user-master.component.css']
+  styleUrls: ['./user-master.component.css'],
+  providers: [ConfirmationService]
 })
 export class UserMasterComponent {
   users: any[] = [];
@@ -17,7 +21,8 @@ export class UserMasterComponent {
   constructor(private userService: UserService,
     private UserManagementService: UserManagement,
     private dialogService: DialogService,
-    private Router: Router) { }
+    private Router: Router,
+    private ConfirmationService:ConfirmationService ) { }
 
   ngOnInit() {
 
@@ -32,24 +37,41 @@ export class UserMasterComponent {
   }
 
 
-  deleteUser(user: any) {
-    this.UserManagementService.DeleteUser(user).subscribe(res => {
-      const ref = this.dialogService.open(DialogeComponent, {
-        header: 'Information',
-        style: {
-          'min-width': '500px',
-          'min-height': '200px',
-        },
-        data: {
-          message: 'Record Deleted Succesfully.',
-        },
+      deleteUser(user: any,event:Event) {
+        this.ConfirmationService.confirm({
+          target: event.target as EventTarget,
+          message: 'Do you want to delete this record?',
+          header: 'Delete Confirmation',
+          icon: 'pi pi-info-circle',
+          acceptButtonStyleClass:"p-button-danger p-button-text",
+          rejectButtonStyleClass:"p-button-text p-button-text",
+          acceptIcon:"none",
+          rejectIcon:"none",
+    
+          accept: () => {
+              // this.messageService.add({ severity: 'info', summary: 'Confirmed', detail: 'Record deleted' });
+              this.UserManagementService.DeleteUser(user).subscribe(res => {
+              const ref = this.dialogService.open(DialogeComponent, {
+                header: 'Information',
+              
+                data: {
+                  message: 'Record updated Succesfully.',
+                },
+              });
+              ref.onClose.subscribe((confirmed: boolean) => {
+               
+                window.location.href = '/UserMaster';
+              })
+            })
+           
+          },
+       
+          reject: () => {
+              // this.messageService.add({ severity: 'error', summary: 'Rejected', detail: 'You have rejected' });
+          }
       });
-      ref.onClose.subscribe((confirmed: boolean) => {
-        window.location.href = '/UserMaster';
-      })
-    })
-  }
-
+      }
+    
 
 
   GetUser(isDeleted: any) {
@@ -61,25 +83,44 @@ export class UserMasterComponent {
     )
   }
 
-  onUnableButtonClick(user: any) {
-    this.UserManagementService.EnableUser(user).subscribe(res => {
-      const ref = this.dialogService.open(DialogeComponent, {
-        header: 'Information',
-        style: {
-          'min-width': '500px',
-          'min-height': '200px', // Adjust the height as needed
+  onUnableButtonClick(user: any,event:Event) {
+
+  
+   this.ConfirmationService.confirm({
+          target: event.target as EventTarget,
+          message: 'Do you want to Enable this record?',
+          header: 'Delete Confirmation',
+          icon: 'pi pi-info-circle',
+          acceptButtonStyleClass:"p-button-danger p-button-text",
+          rejectButtonStyleClass:"p-button-text p-button-text",
+          acceptIcon:"none",
+          rejectIcon:"none",
+          accept: () => {
+            // this.messageService.add({ severity: 'info', summary: 'Confirmed', detail: 'Record deleted' });
+            this.UserManagementService.EnableUser(user).subscribe(res => {
+            const ref = this.dialogService.open(DialogeComponent, {
+              header: 'Information',
+            
+              data: {
+                message: 'Record updated Succesfully.',
+              },
+            });
+            ref.onClose.subscribe((confirmed: boolean) => {
+             
+              window.location.href = '/UserMaster';
+            })
+          })
+         
         },
-        data: {
-          message: 'Record Enabled Succesfully.',
-        },
+     
+        reject: () => {
+            // this.messageService.add({ severity: 'error', summary: 'Rejected', detail: 'You have rejected' });
+        }
       });
-      ref.onClose.subscribe((confirmed: boolean) => {
-        window.location.href = '/UserMaster';
-      })
-      this.ngOnInit()
-    }
-    )
   }
+  
+      
+    
 
   retrieveData(): void {
     if (this.isChecked) {
